* re.c (match_begin): should return offset by character.   [ruby-dev:32331]
* re.c (match_end): ditto.

* re.c (rb_reg_search): ditto.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@13999 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ include/ruby/re.h
@@ -37,9 +37,9 @@ struct RMatch {
 #define RMATCH(obj)  (R_CAST(RMatch)(obj))
 
 VALUE rb_reg_regcomp(VALUE);
-long rb_reg_search(VALUE, VALUE, long, long);
+int rb_reg_search(VALUE, VALUE, int, int);
 VALUE rb_reg_regsub(VALUE, VALUE, struct re_registers *, VALUE);
-long rb_reg_adjust_startpos(VALUE, VALUE, long, long);
+int rb_reg_adjust_startpos(VALUE, VALUE, int, int);
 void rb_match_busy(VALUE);
 VALUE rb_reg_quote(VALUE);
